% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filters.R
\name{apply_filters}
\alias{apply_filters}
\title{Manual data filtering ahead of plotting.}
\usage{
apply_filters(processed, filters)
}
\arguments{
\item{processed}{A list of processed data frames.}

\item{filters}{A list of filter settings.}
}
\value{
A list of filtered data frames.
}
\description{
Manual data filtering ahead of plotting.
}
\examples{
processed <- list(
  general_info = data.frame(patno = c(1, 2), outcome = c("Relapse", "Remission")),
  treatment = data.frame(patno = 1, treatment = "Azacitidine"),
  mrd = data.frame(patno = 1, level = 0.2),
  immune = data.frame(patno = 1),
  ngs = data.frame(patno = 1, Gen = "TP53")
)
filters <- list(outcomes = "Relapse")
apply_filters(processed, filters)
}
