% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{class_finder}
\alias{class_finder}
\title{Called by the draw_clinical_course function to
list the class of each element in the processed data list.}
\usage{
class_finder(x)
}
\arguments{
\item{x}{An object whose class needs to be identified.}
}
\value{
A string indicating the class of the object.
}
\description{
Called by the draw_clinical_course function to
list the class of each element in the processed data list.
}
\examples{
class_finder(processed)
}
