% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{column_check}
\alias{column_check}
\title{Called by preprocess_data to check that all
required columns are present in the provided file.}
\usage{
column_check(df = NULL, required_columns = NULL)
}
\arguments{
\item{df}{Data frame with columns you want to check.}

\item{required_columns}{List of required columns.}
}
\value{
Stops the function and raises a message if columns are missing, else passes.
}
\description{
Called by preprocess_data to check that all
required columns are present in the provided file.
}
\examples{
\dontrun{
column_check(dli_raw, c("patno", "dlidat"))
}
}
