% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_end_date_df}
\alias{create_end_date_df}
\title{Creates a gvhd data frame based on general_info_raw and mrd_raw,
that contains patno, transpldt and rel_term_dat}
\usage{
create_end_date_df(general_info_raw = NULL, mrd_raw = NULL)
}
\arguments{
\item{general_info_raw}{Data frame containing patno,
termindat and transpldt columns.}

\item{mrd_raw}{Data frame containing patno and MRDdat columns.}
}
\value{
Data frame containing patno, transpldt
and rel_term dat columns.
}
\description{
Creates a gvhd data frame based on general_info_raw and mrd_raw,
that contains patno, transpldt and rel_term_dat
}
\examples{
\dontrun{
create_end_date_df(general_info_raw, mrd_raw)
}
}
