% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_gvhd_df}
\alias{create_gvhd_df}
\title{Creates a gvhd data frame based on gvhd_raw and end_date_df, that contains
gvhd events occuring at and in between regular checkups.}
\usage{
create_gvhd_df(gvhd_raw = NULL, end_date_df = NULL)
}
\arguments{
\item{gvhd_raw}{Data frame containing patno, agvhdstage, gvhddate,
agvhdmaxstage, agvhdmaxdt, cgvhdstage, cgvhdmaxstage, cgvhdmaxdt columns.}
}
\value{
Data frame containing patno, gvhd, rel_gvhd_dat,
agvhdstage, cgvhdstage columns.
}
\description{
Creates a gvhd data frame based on gvhd_raw and end_date_df, that contains
gvhd events occuring at and in between regular checkups.
}
