% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_treatment_df}
\alias{create_treatment_df}
\title{Creates treatment data frame based on dli and aza data frames.}
\usage{
create_treatment_df(aza_raw = NULL, dli_raw = NULL, end_date_df = NULL)
}
\arguments{
\item{aza_raw}{Data frame containing patno,
azacitstart and transpldt columns.}

\item{dli_raw}{Data frame containing patno, dlidat and transpldt columns.}

\item{end_date_df}{Data frame containing patno, transpldt
and rel_term_dat columns.}
}
\value{
Data frame containing azacitidine and dli treatment timepoints.
}
\description{
Creates treatment data frame based on dli and aza data frames.
}
\examples{
create_treatment_df(aza_raw, dli_raw, end_date_df)
}
