% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_clinical_course.R
\name{draw_mrd_plot}
\alias{draw_mrd_plot}
\title{Generate MRD plot for a given patient}
\usage{
draw_mrd_plot(mrd_data, general_info_data, ngs_data, x_range, y_upper, pat_id)
}
\arguments{
\item{x_range}{The range of the x-axis.}

\item{y_upper}{The upper limit of the y-axis.}

\item{data}{A data frame containing longitudinal MRD data}
}
\value{
A ggplot object representing the MRD plot for the patient.
}
\description{
Generate MRD plot for a given patient
}
\examples{
draw_mrd_plot(d$mrd)
}
