% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interval_finder}
\alias{interval_finder}
\title{Called by the preprocess_data function to identify
immune suppression intervals, based on the immune suppression data frame.}
\usage{
interval_finder(df)
}
\arguments{
\item{df}{Immune suppression data frame.}
}
\value{
A data frame with identified immune suppression intervals.
}
\description{
Called by the preprocess_data function to identify
immune suppression intervals, based on the immune suppression data frame.
}
\examples{
interval_finder(immune)
}
