% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_dummy_legend}
\alias{make_dummy_legend}
\title{Called by the draw_clinical_course function to create GVHD dummy legends.}
\usage{
make_dummy_legend(levels, colours, title)
}
\arguments{
\item{levels}{The levels of the legend.}

\item{colours}{The colours for the legend.}

\item{title}{The title of the legend.}
}
\value{
A legend grob.
}
\description{
Called by the draw_clinical_course function to create GVHD dummy legends.
}
\examples{
make_dummy_legend(c("0", "1", "2"), c("red", "blue", "green"), "GVHD Stage")
}
