% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data ahead of manual filtering and plotting and.}
\usage{
preprocess_data(
  general_info_file,
  mrd_file,
  dli_file,
  aza_file,
  immune_file,
  gvhd_file,
  ngs_file,
  immune_filter_file,
  chimerism_file,
  strata_filename = list("general_info", "aza", "dli", "ngs", "chimerism", "mrd",
    "immune"),
  strata_colname = NULL,
  strata_itemname = NULL
)
}
\arguments{
\item{general_info_file}{Path to the general info Excel file.}

\item{mrd_file}{Path to the MRD Excel file.}

\item{dli_file}{Path to the DLI Excel file.}

\item{aza_file}{Path to the Azacitidine Excel file.}

\item{immune_file}{Path to the immune suppression Excel file.}

\item{gvhd_file}{Path to the GVHD events Excel file.}

\item{ngs_file}{Path to the NGS Excel file.}

\item{immune_filter_file}{Path to the immune suppression filter CSV file.}

\item{chimerism_file}{Path to the chimerism Excel file.}

\item{strata_filename}{Name of the file containing the strata of interest.}

\item{strata_colname}{Name of the column containing the strata of interest.}

\item{strata_itemname}{Name of the item of the strata of interest.}
}
\value{
A list of processed data frames.
}
\description{
Preprocess data ahead of manual filtering and plotting and.
}
\examples{
preprocess_data(
general_info_file = "general_info.xlsx",
mrd_file = "mrd.xlsx",
dli_file = "dli.xlsx",
aza_file = "aza.xlsx",
immune_file = "immune.xlsx",
gvhd_file = "gvhd.xlsx",
ngs_file = "ngs.xlsx",
immune_filter_file = "immune_filter.csv"
)
}
