% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_one_patient}
\alias{select_one_patient}
\title{Called by the draw_clinical_course function to select one patient per graph.}
\usage{
select_one_patient(df, pat_id = NULL)
}
\arguments{
\item{df}{Data frame containing patient information with a "patno" column.}
}
\value{
Filtered data frame for a specific patient,
or the original data frame if it does not meet the criteria.
}
\description{
Called by the draw_clinical_course function to select one patient per graph.
}
\examples{
select_one_patient(df)
}
