% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_drug}
\alias{standardize_drug}
\title{Called by the preprocess_data function to standardize
drug names in the immune suppression data frame,
based on a provided mapping data frame.}
\usage{
standardize_drug(drug, mapping_df)
}
\arguments{
\item{drug}{Drug name to standardize.}

\item{mapping_df}{Mapping data frame with patterns and standardized names.}
}
\value{
The standardized drug name or NA if no match is found.
}
\description{
Called by the preprocess_data function to standardize
drug names in the immune suppression data frame,
based on a provided mapping data frame.
}
\examples{
\dontrun{
standardize_drug("Drug A", immune_suppression_filter)
}
}
