% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{x_range_finder}
\alias{x_range_finder}
\title{Called by the draw_clinical_course function to
find the range of MRD x-axis values.}
\usage{
x_range_finder(general_info_data)
}
\arguments{
\item{general_info_data}{A data frame containing the
termination date of the patient, relative to the transplantation date.}
}
\value{
A numeric vector indicating the range of the x-axis.
}
\description{
Called by the draw_clinical_course function to
find the range of MRD x-axis values.
}
\examples{
\dontrun{
x_range_finder(d$general_info)
}
}
