% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_chimerism.R
\name{draw_chimerism_plot}
\alias{draw_chimerism_plot}
\title{Generate MRD + chimerism plot for a given patient.}
\usage{
draw_chimerism_plot(
  mrd_data,
  general_info_data,
  ngs_data = NULL,
  chimerism_data,
  x_range,
  y_upper,
  pat_id
)
}
\arguments{
\item{mrd_data}{Mandatory argument:
a data frame containing longitudinal MRD data.}

\item{general_info_data}{Mandatory argument:
a data frame containing general patient info data.}

\item{ngs_data}{A data frame containing ngs data.}

\item{chimerism_data}{A data frame containing chimerism data.}

\item{x_range}{The range of the x-axis.}

\item{y_upper}{The upper limit of the y-axis.}

\item{pat_id}{The patient id of the patient you want to plot.}
}
\value{
A ggplot object representing the MRD plot for the patient.
}
\description{
Generate MRD + chimerism plot for a given patient.
}
\examples{
draw_chimerism_plot(
  d$mrd,
  d$general_info,
  d$ngs_data,
  d$chimerism_data,
  x_range,
  y_upper,
  pat_id
)
}
