% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_chimerism.R
\name{draw_chimerism_plot}
\alias{draw_chimerism_plot}
\title{Generate chimerism plot for a given patient}
\usage{
draw_chimerism_plot(chimerism_data, general_info_data, x_range, pat_id)
}
\arguments{
\item{x_range}{The range of the x-axis.}

\item{data}{A data frame containing longitudinal MRD data}

\item{y_upper}{The upper limit of the y-axis.}
}
\value{
A ggplot object representing the MRD plot for the patient.
}
\description{
Generate chimerism plot for a given patient
}
\examples{
draw_mrd_plot(d$mrd)
}
