% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_clinical_course.R
\name{draw_events_plot}
\alias{draw_events_plot}
\title{Generate events plot for a given patient}
\usage{
draw_events_plot(gvhd_data, immune_intervals_data, treatment_data, x_range)
}
\arguments{
\item{gvhd_data}{A data frame containing longitudinal GVHD data}

\item{immune_intervals_data}{A data frame containing
immune suppression intervals}

\item{treatment_data}{A data frame containing treatment data}

\item{x_range}{The range of the x-axis.}
}
\value{
A ggplot object representing the events plot for the patient.
}
\description{
Generate events plot for a given patient
}
\examples{
draw_events_plot(d$gvhd, d$immune_intervals, d$treatment, x_range)
}
