% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_swimmerplot.R
\name{draw_swimmerplot}
\alias{draw_swimmerplot}
\title{Draw swimmer plot and export to PNG.}
\usage{
draw_swimmerplot(
  processed,
  patient_subset = NULL,
  output_folder,
  output_filename = "swimmerplot.svg",
  output_format = c("svg", "pdf")
)
}
\arguments{
\item{processed}{A list of data frames containing the processed data
for general_info, treatment, mrd, gvhd, immune_intervals, and ngs.}

\item{patient_subset}{A vector of patient IDs.}

\item{output_folder}{The folder where the output PNG will be saved.}

\item{output_filename}{The name of the output PNG file.}
}
\value{
A numeric vector.
}
\description{
Draw swimmer plot and export to PNG.
}
\examples{
draw_swimmerplot(processed, patient_subset, "~/output", "swimmerplot.png")
}
