% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nmds_figures_main}
\alias{nmds_figures_main}
\title{Run analysis using the nmdsFiguresMain function}
\usage{
nmds_figures_main(
  general_info_file,
  mrd_file,
  dli_file,
  aza_file,
  immune_file,
  gvhd_file,
  ngs_file,
  immune_filter_file,
  output_folder,
  plot_type = c("swimmerplot", "clinical_course"),
  filters = NULL
)
}
\arguments{
\item{general_info_file}{Excel file with general patient clinical information}

\item{mrd_file}{Excel file with MRD measurements}

\item{dli_file}{Excel file with DLI treatment information}

\item{aza_file}{Excel file with Azacitidine treatment information}

\item{immune_file}{Excel file with immune suppression treatment information}

\item{gvhd_file}{Excel file with GVHD events information}

\item{ngs_file}{Excel file with NGS information}

\item{immune_filter_file}{CSV file with immune suppression filter}

\item{output_folder}{Folder for generated figures}

\item{plot_type}{Either "swimmerplot" or "clinical_course"}

\item{filters}{Optional filter list or path to JSON file}
}
\description{
Run analysis using the nmdsFiguresMain function
}
\examples{
nmds_figures_main(
general_info_file = "~/general.xlsx",
mrd_file = "~/mrd.xlsx",
dli_file = "~/dli.xlsx",
aza_file = "~/aza.xlsx",
immune_file = "~/immune.xlsx",
gvhd_file = "~/gvhd.xlsx",
ngs_file = "~/ngs.xlsx",
immune_filter_file = "~/immune_filter.csv",
output_folder = "~/output",
plot_type = "clinical_course",
filters = NULL
)
}
