% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nmds_figures_main}
\alias{nmds_figures_main}
\title{Run analysis using the nmds_figures_main function.}
\usage{
nmds_figures_main(
  general_info_file,
  mrd_file,
  dli_file = NULL,
  aza_file = NULL,
  immune_file = NULL,
  gvhd_file = NULL,
  ngs_file = NULL,
  chimerism_file = NULL,
  immune_filter_file = NULL,
  strata_filename = NULL,
  strata_colname = NULL,
  strata_itemname = NULL,
  survival_baseline = NULL,
  survival_metric = NULL,
  output_folder,
  plot_type = c("swimmerplot", "clinical_course", "clinical_course_chimerism",
    "survival"),
  filters = NULL,
  output_format = c("svg", "pdf")
)
}
\arguments{
\item{general_info_file}{Excel file with general patient clinical information.
Required for all plot types.}

\item{mrd_file}{Excel file with MRD measurements. Required for all plot types.}

\item{dli_file}{Excel file with DLI treatment information}

\item{aza_file}{Excel file with Azacitidine treatment information}

\item{immune_file}{Excel file with immune suppression treatment information}

\item{gvhd_file}{Excel file with GVHD events information}

\item{ngs_file}{Excel file with NGS information}

\item{chimerism_file}{Excel file with chimerism information.
Required when \code{plot_type = "clinical_course_chimerism"}.}

\item{immune_filter_file}{CSV file with immune suppression filter}

\item{strata_filename}{Optional file name or data source used for strata.}

\item{strata_colname}{Optional column name for strata.}

\item{strata_itemname}{Optional strata value to keep.}

\item{survival_baseline}{Baseline for survival analysis, one of "transplant", "rel_pos_mrd_dat_0.1", "rel_pos_mrd_dat_0.5", or "rel_pos_mrd_dat_1.0".}

\item{survival_metric}{Survival metric, one of "os", "rfs", or "efs".}

\item{output_folder}{Folder for generated figures}

\item{plot_type}{Either "swimmerplot", "clinical_course", "clinical_course_chimerism", or "survival"}

\item{filters}{Optional filter list}

\item{output_format}{Output format, either "svg" or "pdf"}
}
\description{
Run analysis using the nmds_figures_main function.
}
\examples{
nmds_figures_main(
 general_info_file = "general.xlsx",
 mrd_file = "mrd.xlsx",
 dli_file = "dli.xlsx",
 aza_file = "aza.xlsx",
 immune_file = "immune.xlsx",
 gvhd_file = "gvhd.xlsx",
 ngs_file = "ngs.xlsx",
 chimerism_file = "chimerism.xlsx",
 immune_filter_file = "immune_filter.csv",
 output_folder = "output",
 plot_type = "clinical_course",
 filters = list(
   genes = "TP53",
   outcomes = "Relapse"
 )
)
}
