% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_clinical_course.R
\name{plot_patient_timeline}
\alias{plot_patient_timeline}
\title{Generate MRD + GVHD timeline for a given patient}
\usage{
plot_patient_timeline(processed, pat_id)
}
\arguments{
\item{processed}{A list of data frames containing the processed data
for general_info, treatment, mrd, gvhd, immune_intervals, and ngs.}

\item{pat_id}{A vector of patient IDs.}
}
\value{
A numeric vector.
}
\description{
Generate MRD + GVHD timeline for a given patient
}
\examples{
plot_patient_timeline(processed, pat_id)
}
