% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_swimmerplot.R
\name{swimmerplot}
\alias{swimmerplot}
\title{Called by draw_swimmerplot to draw swimmer plot.}
\usage{
swimmerplot(
  plot_data,
  immune_pts,
  outcome_pts,
  treatment_pts,
  gvhd_pts,
  mrd_terminal_pts,
  title_string
)
}
\arguments{
\item{plot_data}{Data frame containing
the data for the swimmer plot rectangles.}

\item{immune_pts}{A vector of patient IDs.}

\item{outcome_pts}{A vector of patient IDs.}

\item{treatment_pts}{A vector of patient IDs.}

\item{gvhd_pts}{A vector of patient IDs.}

\item{mrd_terminal_pts}{A data frame describing mrd values taken
at the termination date, with the columns patno, rel_term_dat,
y and mrd_category.}

\item{title_string}{The title for the swimmer plot.}
}
\value{
A swimmer plot ggplot object.
}
\description{
Called by draw_swimmerplot to draw swimmer plot.
}
\examples{
swimmerplot(
plot_data,
immune_pts,
outcome_pts,
treatment_pts,
gvhd_pts,
mrd_terminal_pts,
title_string
)
}
